cmd/lava: rename "lava run" command to "lava scan"
This commit releases the "lava run" command, which we may want to use
in the future to compile+run local Vulcan checks that are under
development.
